Have Mesh create it's own GUI
