+++ src/Scene.h
@@ -1,5 +1,4 @@
 #pragma once
-#include"Mesh.h"
 #include<vector>
 #include<glm/glm.hpp>
 #include"Renderer.h"
